refactor: forms fix: form validate
