+++ src/components/Main/Auth/Auth.module.scss
@@ -32,7 +32,7 @@
         h3 {
           text-align: left;
         }
-        p {
+        >p {
           font: $typoDescribeForms;
           text-align: left;
           color: $secondary;
