+++ src/components/Main/ChangePass/ChangePass/ChangePass.module.scss
@@ -1,5 +1,5 @@
-@import './src/styles/mixins.module';
-@import './src/styles/variables.module';
+@import '../../../../styles/mixins.module';
+@import '../../../../styles/variables.module';
 
 .container {
     @include gridContainer (100.8rem, transparent);
